pyproject.toml
[project]
name = "text-to-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.39.0",
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-openai>=0.3.28",
    "openai>=1.98.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pyprojroot>=0.3.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.42",
]
